BREAKING: show current app in AppMenu and remove unused ConnectPeople filtering
The AppMenu always shows the current app.

The ConnectedPeople menu no longer attempts to filter out current user.
It was supposed to but never did because I was filtering on `name` instead of `organization_name`.
Nobody cared so I'd hate to break it now.
Especially given the behaviour change to AppMenu.

* https://trello.com/c/IpTjGUWh/2746-453680-when-i-click-on-upper-left-services-nothing-happens