spike, day 1
Danny D and I took 90 minutes today to spike what skeletons might look
like for Doxy Elements.

We chose a subset of elements that make sense for congregant-facing
features, which includes a lot of media elements and no form elements.

We've included notes, in code, for next step and examples to build up.