change(ApiRequest): assume fetch, but support injection
An existing expectation of topbar is that it will be able to make
authenticated requests to the Planning Center API. While yes, data
_could_ be passed in, these data providers are designed to speak to
the API automatically.

This commit moves topbar away from `superagent`, and towards `fetch`,
with the ability to inject a custom function for talking to the
API. The new contract with a host app is that when you mount the
topbar, you will also tell topbar how to make authenticated reqeusts
to the Planning Center API. For a Rails app, this will mostly mean
adding `CSRF` header, but for an SPA, it would mean adding an
`Authentication` header. It's flexible for the needs of the environment.

The built in `apiRequest` method follows `fetch` closely, with added
`data` and `onError` options, to fit well within the existing Topbar
design. As it stands today, this should work for all `GET` requests
when mounted in a Rails app (with session authentication), but
wouldn't work for a `POST` (etc) request, because of the need for a
CSRF token. You can see that removed from the `platform_announcements`
component.