change(Topbar::ApiRequest): affordances for Rails apps
We want it to be _possible_ to mount Topbar in any app, but we
recognize that right now we only mount it in Rails apps, and we can
automatically infer how to properly make authenticated API requests in
that context.

The `defaultFetch` function now expreses that, automatically detecting
the presence of a Rails CSRF token, and setting it as a header
automatically.

Host apps are still able to inject their own fetch if they need to,
but it's no longer required for our golden path use case.