[Draft] use Login for the top right corner (#46)
* change(ApiRequest): assume fetch, but support injection

An existing expectation of topbar is that it will be able to make
authenticated requests to the Planning Center API. While yes, data
_could_ be passed in, these data providers are designed to speak to
the API automatically.

This commit moves topbar away from `superagent`, and towards `fetch`,
with the ability to inject a custom function for talking to the
API. The new contract with a host app is that when you mount the
topbar, you will also tell topbar how to make authenticated reqeusts
to the Planning Center API. For a Rails app, this will mostly mean
adding `CSRF` header, but for an SPA, it would mean adding an
`Authentication` header. It's flexible for the needs of the environment.

The built in `apiRequest` method follows `fetch` closely, with added
`data` and `onError` options, to fit well within the existing Topbar
design. As it stands today, this should work for all `GET` requests
when mounted in a Rails app (with session authentication), but
wouldn't work for a `POST` (etc) request, because of the need for a
CSRF token. You can see that removed from the `platform_announcements`
component.

* feat(Topbar::ConnectedPeople): add `unlink` action

Unlinking, historically, has been a `GET` request, and it's time to
switch to a verb.

To keep much of Topbar's existing design and goals in tact, we'll
accomplish this through the API.

# TODO: add the link to the Login PR

* feat(Topbar::ConnectedPeople): add switch

Just like we did with Unlink, there's a similar action in Login for
switching accounts. It has the same interaction of `POST`-ing, and
getting a new redirect path afterwards.

* change(Topbar::ConnectedPeople): use Login for this info

Login is adding this edge, and since it controls linking and
unlinking, it's a natural fit to look there for this info

* change(Topbar/ApiRequest): ask for an authenticated fetch

Rather than requiring an API request function that fits our
convenience methods, just ask for something that has the same in/out
signature as `window.fetch`, but makes an authenticated request.

A working example is here in Giving: https://github.com/ministrycentered/giving/pull/4482

This keeps the pros of the convenience that `apiRequest`
offers (`data` for JSON, automatic error handling, and automatic JSON
parsing), while keeping the public API very `fetch` focused.

* change(Topbar::ApiRequest): affordances for Rails apps

We want it to be _possible_ to mount Topbar in any app, but we
recognize that right now we only mount it in Rails apps, and we can
automatically infer how to properly make authenticated API requests in
that context.

The `defaultFetch` function now expreses that, automatically detecting
the presence of a Rails CSRF token, and setting it as a header
automatically.

Host apps are still able to inject their own fetch if they need to,
but it's no longer required for our golden path use case.

* rename(Topbar): authenticatedFetch -> configuredFetch

Move to a more generic name to express that host apps can inject
whatever `fetch` they need, that's configured appropriately to make
requests in their context. Today the need is authentication, but
`configure` leaves the door open for needs that appear in the future.

* change(Topbar/ApiRequest): export `railsCsrfFetch`

The default `fetch` is a sane default, now, and moving forward.

We also export a Rails-friendly `fetch` that can be used _right now_
with our existing apps. All they have to do is pass this function
app-side.