chore(topbar): extract Topbar into its own repo
While lerna seemed like a good idea in the beginning of these
projects, it is starting to cause more problems than it was intended
to solve. Topbar in particular is soon going to have some major
features introduced and extracting it from being tied to all the other
repos in this monorepo will make life easier for any future dev work.

The new repo can be found at ministrycentered/topbar.

I've done my best to maintain the commit history that deals directly
with topbar including transfering every tag from 4.0.0–7.0.0. For
future context, this was done with the help of [git-filter-repo][gfr].

[gfr]: https://github.com/newren/git-filter-repo